feat(stateful-nat): move NF to nf.rs
The stateful NAT grew significantly. Move the source implementation
of the NF to nf.rs
